CoverSystem in action
After coverRequest, bots will find the nearest cover and use it. When there is no free coverSlot available, bots will continue on what they were doing. 
When a coverSlot frees up, a bot will use the slot if it still has the command to take cover. 

Made with Unity DOTS